fix msvc compiler warning about possible size_t/int value truncation
Warning reported when you attempt to implicitly convert between values of types
int & size_t as those types may be of different sized and int is signed while
size_t is not.

+++ source/bases/Win32GUI.c
@@ -255,7 +255,7 @@ int WINAPI WinMain(
 #if PY_MAJOR_VERSION >= 3
     char fileName[MAXPATHLEN + 1];
     wchar_t **wargv, *wfileName;
-    int i, size;
+    size_t i, size;
 #else
     const char *fileName;
 #endif
